added another component and edit to readme
