+++ README.md
@@ -1,2 +1,2 @@
-# epa-practice
+# Avalanche Wiki
 Making a lightweight streamlit app that takes in a user's zipcode and returns the latest EPA data about the areas air quality
